Consolidating some overlapping API call implementations.
